refactor: move entity error classes into the errors domain
EntityRelationLookupError and the new EntityMetadataError now live in
src/errors as standalone files, one class per file, alongside the rest
of the TypeormExtensionError hierarchy instead of next to their thrower.
