+++ src/errors/entity-metadata.ts
@@ -1,0 +1,11 @@
+import { TypeormExtensionError } from './base';
+
+export class EntityMetadataError extends TypeormExtensionError {
+    constructor(message?: string) {
+        super(message || 'An entity metadata related error has occurred');
+    }
+
+    static notRegistered(entity: unknown) {
+        return new EntityMetadataError(`The entity ${entity} is not registered.`);
+    }
+}
