+++ src/errors/index.ts
@@ -1,5 +1,7 @@
 export * from './base';
 export * from './driver';
+export * from './entity-metadata';
+export * from './entity-relation-lookup';
 export * from './options';
 export * from './schema';
 export * from './schema-alteration';
